+++ packages/ethereum/sdk/src/common/balances.test.ts
@@ -21,6 +21,12 @@ describe("getBalance test", () => {
 
 	const testErc20Address = toAddress("0xa03C1eCaEB1D8A7581FC38d28f67c3d42a8B9b76")
 
+	test.concurrent("get eth balance", async () => {
+		const senderAddress = toAddress("0xC072c9889dE7206c1C18B9d9973B06B8646FC6bd")
+		const balance = await balances.getBalance(senderAddress, { assetClass: "ETH" })
+		expect(balance.toString()).toBe("0")
+	})
+
 	test.concurrent("get non-zero eth balance", async () => {
 		const senderAddress = toAddress("0xa14FC5C72222FAce8A1BcFb416aE2571fA1a7a91")
 		const balance = await balances.getBalance(senderAddress, { assetClass: "ETH" })