+++ packages/ethereum/sdk/src/common/test/test-credentials.ts
@@ -60,6 +60,7 @@ export function getAPIKey(env:  EthereumNetwork) {
 		case "mantle":
 		case "polygon":
 		case "arbitrum":
+		case "rari":
 		case "zksync":
 		case "chiliz":
 		case "lightlink":