+++ packages/ethereum/sdk/src/nft/burn.test.ts
@@ -23,6 +23,7 @@ import { burn as burnTemplate } from "./burn"
 import { ERC1155VersionEnum, ERC721VersionEnum } from "./contracts/domain"
 import { getErc721Contract } from "./contracts/erc721"
 import { getErc1155Contract } from "./contracts/erc1155"
+import { awaitOwnership } from "./test/await-ownership"
 
 const { provider, wallet } = createE2eProvider(DEV_PK_1)
 const { providers } = createTestProviders(provider, wallet)
@@ -56,6 +57,7 @@ describe.each(providers)("burn nfts", (ethereum: Ethereum) => {
 		if (minted.type === MintResponseTypeEnum.ON_CHAIN) {
 			await minted.transaction.wait()
 		}
+		await awaitOwnership(await getApis(), e2eErc721V2ContractAddress, minted.tokenId, testAddress)
 		const testBalance = await testErc721.functionCall("balanceOf", testAddress).call()
 
 		const burnTx = await burn({
@@ -83,6 +85,7 @@ describe.each(providers)("burn nfts", (ethereum: Ethereum) => {
 		if (minted.type === MintResponseTypeEnum.ON_CHAIN) {
 			await minted.transaction.wait()
 		}
+		await awaitOwnership(await getApis(), e2eErc1155V1ContractAddress, minted.tokenId, testAddress)
 		const burnTx = await burn({
 			assetType: {
 				contract: e2eErc1155V1ContractAddress,
@@ -109,6 +112,7 @@ describe.each(providers)("burn nfts", (ethereum: Ethereum) => {
 		if (minted.type === MintResponseTypeEnum.ON_CHAIN) {
 			await minted.transaction.wait()
 		}
+		await awaitOwnership(await getApis(), e2eErc721V3ContractAddress, minted.tokenId, testAddress)
 		const tx = await burn({
 			assetType: {
 				contract: e2eErc721V3ContractAddress,
@@ -140,6 +144,7 @@ describe.each(providers)("burn nfts", (ethereum: Ethereum) => {
 		if (minted.type === MintResponseTypeEnum.ON_CHAIN) {
 			await minted.transaction.wait()
 		}
+		await awaitOwnership(await getApis(), e2eErc1155V2ContractAddress, minted.tokenId, testAddress)
 		const tx = await burn({
 			assetType: {
 				contract: e2eErc1155V2ContractAddress,
@@ -173,6 +178,7 @@ describe.each(providers)("burn nfts", (ethereum: Ethereum) => {
 		if (minted.type === MintResponseTypeEnum.ON_CHAIN) {
 			await minted.transaction.wait()
 		}
+		await awaitOwnership(await getApis(), e2eErc1155V2ContractAddress, minted.tokenId, testAddress)
 		const tx = await burn({
 			assetType: {
 				contract: e2eErc1155V2ContractAddress,