+++ packages/sdk/src/config/domain.ts
@@ -1,8 +1,8 @@
-import type { FlowEnv } from "@rarible/flow-sdk/build/types"
 import type { EthereumNetwork } from "@rarible/protocol-ethereum-sdk/build/types"
 import type { TezosNetwork } from "@rarible/tezos-sdk"
 import type { SolanaCluster } from "@rarible/solana-sdk"
 import type { ImxEnv } from "@rarible/immutable-wallet"
+import type { FlowEnv } from "@rarible/flow-sdk"
 
 export type RaribleSdkEnvironment = "development" | "testnet" | "prod"