+++ packages/sdk/src/sdk-blockchains/ethereum/balance.ts
@@ -8,20 +8,19 @@ import { Blockchain } from "@rarible/api-client"
 import { Action } from "@rarible/action"
 import type { Maybe } from "@rarible/types/build/maybe"
 import type { EthereumWallet } from "@rarible/sdk-wallet"
+import type * as ApiClient from "@rarible/api-client"
+import { extractBlockchain } from "@rarible/sdk-common"
 import type { ConvertRequest } from "../../types/balances"
 import type {
 	DepositBiddingBalanceRequest,
 	GetBiddingBalanceRequest,
 	WithdrawBiddingBalanceRequest,
 } from "../../types/balances"
-import { getCurrencyAssetType } from "../../common/get-currency-asset-type"
+import { getCurrencyAssetType, getEVMCurrencyId, isErc20, isEth } from "../../common/get-currency-asset-type"
 import type { RequestCurrency } from "../../common/domain"
 import type { IApisSdk } from "../../domain"
-import { extractBlockchain } from "../../common/extract-blockchain"
 import {
 	convertEthereumContractAddress,
-	convertToEthereumAddress,
-	convertToEthereumAssetType,
 	getWalletNetwork,
 	isEVMBlockchain,
 } from "./common"
@@ -38,13 +37,17 @@ export class EthereumBalance {
 	}
 
 	async getBalance(address: UnionAddress, currency: RequestCurrency): Promise<BigNumber> {
-		const assetType = convertToEthereumAssetType(getCurrencyAssetType(currency))
-		if (assetType.assetClass !== "ETH" && assetType.assetClass !== "ERC20") {
+		const type = getCurrencyAssetType(currency)
+		if (!isEth(type) && !isErc20(type)) {
 			throw new Error("Unsupported asset type for getting balance")
 		}
-		const addressRaw = convertToEthereumAddress(address)
-		const value = await this.sdk.balances.getBalance(addressRaw, assetType)
-		return toBn(value)
+		const response = await this.apis.balances.getBalance({
+			currencyId: getEVMCurrencyId(
+				currency as ApiClient.EthErc20AssetType | ApiClient.EthEthereumAssetType
+			),
+			owner: address,
+		})
+		return toBn(response.decimal)
 	}
 
 	async convert(request: ConvertRequest): Promise<IBlockchainTransaction> {