+++ CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.68-fix.2](https://github.com/rarible/sdk/compare/v0.13.68-fix.1...v0.13.68-fix.2) (2024-01-30)
+
+
+### Features
+
+* **PLTF-275:** get evm balances from union api ([#593](https://github.com/rarible/sdk/issues/593)) ([5f51042](https://github.com/rarible/sdk/commit/5f510422c338e06ea10ab94bc98d0b65031626af))
+
+
+
+
+
+## [0.13.68-fix.1](https://github.com/rarible/sdk/compare/v0.13.68-fix.0...v0.13.68-fix.1) (2024-01-30)
+
+
+### Features
+
+* **PLTF-262:** zksync deploy collection fix ([#592](https://github.com/rarible/sdk/issues/592)) ([2d2be69](https://github.com/rarible/sdk/commit/2d2be691c2491080bb5bced8f4db4576893b1d19))
+
+
+
+
+
+## [0.13.68-fix.0](https://github.com/rarible/sdk/compare/v0.13.68...v0.13.68-fix.0) (2024-01-30)
+
+
+### Features
+
+* **PLTF-264:** replace evm apis in union ([#590](https://github.com/rarible/sdk/issues/590)) ([e516db8](https://github.com/rarible/sdk/commit/e516db8386bae0d95c1f9813c092f19573194999))
+
+
+
+
+
 ## [0.13.68](https://github.com/rarible/sdk/compare/v0.13.67...v0.13.68) (2024-01-11)
 
 
@@ -10,6 +43,35 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 * **PLTF-107:** walletConnect chainId ([#588](https://github.com/rarible/sdk/issues/588)) ([0904b99](https://github.com/rarible/sdk/commit/0904b99cb705526424a06aca55ea70f12f93018e))
 
+
+## [0.13.67-fix.2](https://github.com/rarible/sdk/compare/v0.13.67-fix.1...v0.13.67-fix.2) (2024-01-24)
+
+
+### Features
+
+* **PLTF-229:** split setup collections ([9bfab5b](https://github.com/rarible/sdk/commit/9bfab5b82751a8de1fbcb7e2905c50661cc21795))
+
+
+
+
+
+## [0.13.67-fix.1](https://github.com/rarible/sdk/compare/v0.13.67-fix.0...v0.13.67-fix.1) (2024-01-23)
+
+
+### Features
+
+* split setup gamisodes and mattel collections ([113952a](https://github.com/rarible/sdk/commit/113952a09c2a8ba8ac5b3779908632b59158e8cd))
+
+
+
+
+
+## [0.13.67-fix.0](https://github.com/rarible/sdk/compare/v0.13.67...v0.13.67-fix.0) (2024-01-23)
+
+
+### Features
+
+* rari chain mainnet support ([9c6161f](https://github.com/rarible/sdk/commit/9c6161fab1cab1f776459ebb505e67d2874b5f6a))