+++ packages/ethereum/sdk/CHANGELOG.md
@@ -3,12 +3,69 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.68-fix.2](https://github.com/rarible/sdk/compare/v0.13.68-fix.1...v0.13.68-fix.2) (2024-01-30)
+
+**Note:** Version bump only for package @rarible/protocol-ethereum-sdk
+
+
+
+
+
+## [0.13.68-fix.1](https://github.com/rarible/sdk/compare/v0.13.68-fix.0...v0.13.68-fix.1) (2024-01-30)
+
+
+### Features
+
+* **PLTF-262:** zksync deploy collection fix ([#592](https://github.com/rarible/sdk/issues/592)) ([2d2be69](https://github.com/rarible/sdk/commit/2d2be691c2491080bb5bced8f4db4576893b1d19))
+
+
+
+
+
+## [0.13.68-fix.0](https://github.com/rarible/sdk/compare/v0.13.68...v0.13.68-fix.0) (2024-01-30)
+
+
+### Features
+
+* **PLTF-264:** replace evm apis in union ([#590](https://github.com/rarible/sdk/issues/590)) ([e516db8](https://github.com/rarible/sdk/commit/e516db8386bae0d95c1f9813c092f19573194999))
+
+
+
+
+
 ## [0.13.68](https://github.com/rarible/sdk/compare/v0.13.67...v0.13.68) (2024-01-11)
 
 
 ### Bug Fixes
 
 * **PLTF-107:** walletConnect chainId ([#588](https://github.com/rarible/sdk/issues/588)) ([0904b99](https://github.com/rarible/sdk/commit/0904b99cb705526424a06aca55ea70f12f93018e))
+
+
+
+
+
+## [0.13.67-fix.2](https://github.com/rarible/sdk/compare/v0.13.67-fix.1...v0.13.67-fix.2) (2024-01-24)
+
+**Note:** Version bump only for package @rarible/protocol-ethereum-sdk
+
+
+
+
+
+## [0.13.67-fix.1](https://github.com/rarible/sdk/compare/v0.13.67-fix.0...v0.13.67-fix.1) (2024-01-23)
+
+**Note:** Version bump only for package @rarible/protocol-ethereum-sdk
+
+
+
+
+
+## [0.13.67-fix.0](https://github.com/rarible/sdk/compare/v0.13.67...v0.13.67-fix.0) (2024-01-23)
+
+
+### Features
+
+* rari chain mainnet support ([9c6161f](https://github.com/rarible/sdk/commit/9c6161fab1cab1f776459ebb505e67d2874b5f6a))